Circular reference bug fix : getDTOs methods moved from Service classes to Mapper classes.
