+++ src/main/java/com/pcs/service/RuleNameService.java
@@ -2,7 +2,6 @@ package com.pcs.service;
 
 import com.pcs.model.RuleName;
 import com.pcs.repository.RuleNameRepository;
-import com.pcs.web.dto.RuleNameDTO;
 import com.pcs.web.mapper.RuleNameMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -10,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,8 +17,6 @@ public class RuleNameService {
 
     @Autowired
     private RuleNameRepository ruleNameRepository;
-    @Autowired
-    private RuleNameMapper ruleNameMapper;
 
     private static final Logger logger = LogManager.getLogger(RuleNameService.class);
 
@@ -63,15 +59,6 @@ public class RuleNameService {
             logger.error("Unable to find and delete RuleName corresponding to id {}", id);
             throw new IllegalArgumentException("Invalid ruleName id");
         }
-    }
-
-    public List<RuleNameDTO> getRuleNameDTOs() {
-        List<RuleName> ruleNames = getRuleNames();
-        List<RuleNameDTO> ruleNameDTOs = new ArrayList<>();
-        ruleNames.forEach(ruleName -> {
-            ruleNameDTOs.add(ruleNameMapper.toRuleNameDTO(ruleName));
-        });
-        return ruleNameDTOs;
     }
 
 }
