+++ src/main/java/com/pcs/service/TradeService.java
@@ -2,7 +2,6 @@ package com.pcs.service;
 
 import com.pcs.model.Trade;
 import com.pcs.repository.TradeRepository;
-import com.pcs.web.dto.TradeDTO;
 import com.pcs.web.mapper.TradeMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -10,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,8 +17,6 @@ public class TradeService {
 
     @Autowired
     private TradeRepository tradeRepository;
-    @Autowired
-    private TradeMapper tradeMapper;
 
     private static final Logger logger = LogManager.getLogger(TradeService.class);
 
@@ -63,15 +59,6 @@ public class TradeService {
             logger.error("Unable to find and delete trade corresponding to id {}", id);
             throw new IllegalArgumentException("Invalid trade id");
         }
-    }
-
-    public List<TradeDTO> getTradeDTOs() {
-        List<Trade> trades = getTrades();
-        List<TradeDTO> tradeDTOs = new ArrayList<>();
-        trades.forEach(trade -> {
-            tradeDTOs.add(tradeMapper.toTradeDTO(trade));
-        });
-        return tradeDTOs;
     }
 
 }
