+++ src/main/java/com/pcs/service/UserService.java
@@ -2,7 +2,6 @@ package com.pcs.service;
 
 import com.pcs.model.User;
 import com.pcs.repository.UserRepository;
-import com.pcs.web.dto.UserDTO;
 import com.pcs.web.mapper.UserMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -10,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,8 +18,7 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
-    @Autowired
-    private UserMapper userMapper;
+
     private static final Logger logger = LogManager.getLogger(UserService.class);
 
     public List<User> getUsers() {
@@ -63,15 +60,6 @@ public class UserService {
             logger.error("Unable to find and delete user corresponding to id {}", id);
             throw new IllegalArgumentException("Invalid user id");
         }
-    }
-
-    public List<UserDTO> getUserDTOs() {
-        List<User> users = getUsers();
-        List<UserDTO> userDTOs = new ArrayList<>();
-        users.forEach(user -> {
-            userDTOs.add(userMapper.toUserDTO(user));
-        });
-        return userDTOs;
     }
 
 }
