+++ src/main/resources/templates/user/update.html
@@ -28,6 +28,9 @@
 					<input type="text" th:field="*{username}" id="username" placeholder="User Name" class="col-4">
 					<p class="text-danger" th:if="${#fields.hasErrors('username')}" th:errors="*{username}"></p>
 				</div>
+				<div th:if="${#fields.hasGlobalErrors()}">
+					<p th:each="err : ${#fields.globalErrors() }" th:text="${err}" class="error" />
+				</div>
 			</div>
 			<div class="form-group">
 				<label for="password" class="col-sm-2 control-label">Password</label>
@@ -51,10 +54,6 @@
 					<a class="btn btn-danger btn-sm" href="/user/list">Cancel</a>
 					<input class="btn btn-primary btn-sm" type="submit" value="Update User">
 				</div>
-			</div>
-
-			<div th:if="${#fields.hasGlobalErrors()}">
-				<p th:each="err : ${#fields.globalErrors() }" th:text="${err}" class="error" />
 			</div>
 
 		</form>
