+++ src/test/java/com/pcs/service/CurvePointServiceTest.java
@@ -2,7 +2,6 @@ package com.pcs.service;
 
 import com.pcs.model.CurvePoint;
 import com.pcs.repository.CurvePointRepository;
-import com.pcs.web.dto.CurvePointDTO;
 import com.pcs.web.mapper.CurvePointMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -10,8 +9,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -69,28 +66,6 @@ public class CurvePointServiceTest {
                     curvePointService.deleteById(anyInt());
                 }, "IllegalArgumentException was expected");
         assertEquals("Invalid curvePoint id", thrown.getMessage());
-    }
-
-    @Test
-    public void should_get_curvePointDTOs_successfully() throws Exception {
-        //given
-        CurvePoint curvePoint1 = new CurvePoint(1, 2, 2.2, 3.3);
-        CurvePoint curvePoint2 = new CurvePoint(2, 3, 3.3, 4.4);
-        List<CurvePoint> curvePoints = new ArrayList<>();
-        curvePoints.add(curvePoint1);
-        curvePoints.add(curvePoint2);
-        CurvePointDTO curvePointDTO1 = new CurvePointDTO(1, "2", "2.2", "3.3");
-        CurvePointDTO curvePointDTO2 = new CurvePointDTO(2, "3", "3.3", "4.4");
-        List<CurvePointDTO> expectedCurvePointDTOs = new ArrayList<>();
-        expectedCurvePointDTOs.add(curvePointDTO1);
-        expectedCurvePointDTOs.add(curvePointDTO2);
-        when(curvePointService.getCurvePoints()).thenReturn(curvePoints);
-        when(curvePointMapper.toCurvePointDTO(curvePoint1)).thenReturn(curvePointDTO1);
-        when(curvePointMapper.toCurvePointDTO(curvePoint2)).thenReturn(curvePointDTO2);
-        //when
-        List<CurvePointDTO> actualCurvePointDTOs = curvePointService.getCurvePointDTOs();
-        //then
-        assertEquals(expectedCurvePointDTOs, actualCurvePointDTOs);
     }
 
 }
