+++ src/test/java/com/pcs/service/UserServiceTest.java
@@ -2,7 +2,6 @@ package com.pcs.service;
 
 import com.pcs.model.User;
 import com.pcs.repository.UserRepository;
-import com.pcs.web.dto.UserDTO;
 import com.pcs.web.mapper.UserMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -10,8 +9,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -70,32 +67,6 @@ public class UserServiceTest {
                     userService.deleteById(anyInt());
                 }, "IllegalArgumentException was expected");
         assertEquals("Invalid user id", thrown.getMessage());
-    }
-
-    @Test
-    public void should_get_userDTOs_successfully() throws Exception {
-        //given
-        User user1 = new User(1, "username_1", "crypted_password_1","fullname_1",
-                "user");
-        User user2 = new User(2, "username_2", "crypted_password_2","fullname_2",
-                "user");
-        List<User> users = new ArrayList<>();
-        users.add(user1);
-        users.add(user2);
-        UserDTO userDTO1 = new UserDTO(1, "username_1", "crypted_password_1","fullname_1",
-                "user");
-        UserDTO userDTO2 = new UserDTO(2, "username_2", "crypted_password_2","fullname_2",
-                "user");;
-        List<UserDTO> expectedUserDTOs = new ArrayList<>();
-        expectedUserDTOs.add(userDTO1);
-        expectedUserDTOs.add(userDTO2);
-        when(userService.getUsers()).thenReturn(users);
-        when(userMapper.toUserDTO(user1)).thenReturn(userDTO1);
-        when(userMapper.toUserDTO(user2)).thenReturn(userDTO2);
-        //when
-        List<UserDTO> actualUserDTOs = userService.getUserDTOs();
-        //then
-        assertEquals(expectedUserDTOs, actualUserDTOs);
     }
 
 }
